@@ -1,6 +1,6 @@
 ﻿using Microsoft.Extensions.AI;
 
-IChatClient chatClient = 
+IChatClient chatClient =
     new OllamaChatClient(new Uri("http://localhost:11434/"), "phi3:mini");
 
 // Start the conversation with context for the AI model
@@ -17,7 +17,7 @@ while (true)
     Console.WriteLine("AI Response:");
     var response = "";
     await foreach (var item in
-        chatClient.CompleteStreamingAsync(chatHistory))
+        chatClient.GetStreamingResponseAsync(chatHistory))
     {
         Console.Write(item.Text);
         response += item.Text;
