@@ -2,9 +2,9 @@
 using Microsoft.Extensions.Configuration;
 using OpenAI;
 
-var config = new ConfigurationBuilder().AddUserSecrets<Program>().Build();
-string model = config["ModelName"];
-string key = config["OpenAIKey"];
+IConfigurationRoot config = new ConfigurationBuilder().AddUserSecrets<Program>().Build();
+string? model = config["ModelName"];
+string? key = config["OpenAIKey"];
 
 // Create the IChatClient
 IChatClient client =
@@ -19,5 +19,5 @@ string prompt = $"""
 Console.WriteLine($"user >>> {prompt}");
 
 // Submit the prompt and print out the response
-ChatCompletion response = await client.CompleteAsync(prompt, new ChatOptions { MaxOutputTokens = 400 });
+ChatResponse response = await client.GetResponseAsync(prompt, new ChatOptions { MaxOutputTokens = 400 });
 Console.WriteLine($"assistant >>> {response}");
