@@ -4,7 +4,7 @@ using System.Text;
 namespace interpolated_string_handler.Version3
 {
     [InterpolatedStringHandler]
-    public ref struct LogInterpolatedStringHandler
+    public struct LogInterpolatedStringHandler
     {
         // Storage for the built-up string
         StringBuilder builder;
