@@ -98,6 +98,14 @@ public class BasicObjectInitializers
         };
         // </SnippetListInitializerWithNull>
 
+        // <SnippetListInitializerWithSpreadOperator>
+        List<Cat> allCats = [.. cats, .. moreCats];
+        // </SnippetListInitializerWithSpreadOperator>
+
+        // <SnippetListInitializerWithSpreadOperatorAndAdditionalElement>
+        List<Cat> additionalCats = [.. cats, new Cat { Name = "Furrytail", Age = 5 }, .. moreCats];
+        // </SnippetListInitializerWithSpreadOperatorAndAdditionalElement>
+
         // <SnippetDictionaryIndexerInitializer>
         var numbers = new Dictionary<int, string>
         {
@@ -182,15 +190,10 @@ public class InitializationSample
             null
         };
 
+        List<Cat> allCats = [.. cats, new Cat { Name = "Łapka", Age = 5 }, cat, .. moreCats];
+
         // Display results.
-        System.Console.WriteLine(cat.Name);
-
-        foreach (Cat c in cats)
-        {
-            System.Console.WriteLine(c.Name);
-        }
-
-        foreach (Cat? c in moreCats)
+        foreach (Cat? c in allCats)
         {
             if (c != null)
             {
@@ -203,13 +206,14 @@ public class InitializationSample
         }
     }
     // Output:
-    //Fluffy
-    //Sylvester
-    //Whiskers
-    //Sasha
-    //Furrytail
-    //Peaches
-    //List element has null value.
+    // Sylvester
+    // Whiskers
+    // Sasha
+    // Łapka
+    // Fluffy
+    // Furrytail
+    // Peaches
+    // List element has null value.
 }
 // </SnippetFullExample>
 
