@@ -1,14 +1,20 @@
-﻿public class DerivedClassWithFinalizer : BaseClassWithFinalizer
-{
-    // To detect redundant calls
-    private bool _disposedValue;
+﻿using System.Threading;
 
-    ~DerivedClassWithFinalizer() => Dispose(false);
+public class DisposableDerivedWithFinalizer : DisposableBaseWithFinalizer
+{
+    // Detect redundant Dispose() calls in a thread-safe manner.
+    // _isDisposed == 0 means Dispose(bool) has not been called yet.
+    // _isDisposed == 1 means Dispose(bool) has been already called.
+    private int _isDisposed;
+
+    ~DisposableDerivedWithFinalizer() => Dispose(false);
 
     // Protected implementation of Dispose pattern.
     protected override void Dispose(bool disposing)
     {
-        if (!_disposedValue)
+        // In case _isDisposed is 0, atomically set it to 1.
+        // Enter the branch only if the original value is 0.
+        if (Interlocked.CompareExchange(ref _isDisposed, 1, 0) == 0)
         {
             if (disposing)
             {
@@ -17,7 +23,6 @@
 
             // TODO: free unmanaged resources (unmanaged objects) and override a finalizer below.
             // TODO: set large fields to null.
-            _disposedValue = true;
         }
 
         // Call the base class implementation.
