@@ -6,162 +6,6 @@ using System.Collections.Generic;
 namespace CsProgGuideTypes
 {
     //-------------------------------------------------------------------------
-    class WrapCasting
-    {
-        //<Snippet1>
-        static void TestCasting()
-        {
-            int i = 10;
-            float f = 0;
-            f = i;  // An implicit conversion, no data will be lost.
-            f = 0.5F;
-            i = (int)f;  // An explicit conversion. Information will be lost.
-        }
-        //</Snippet1>
-
-        //<Snippet2>
-        class Test
-        {
-            static void Main()
-            {
-                double x = 1234.7;
-                int a;
-                // Cast double to int.
-                a = (int)x;
-                System.Console.WriteLine(a);
-            }
-        }
-        // Output: 1234
-        //</Snippet2>
-    }
-
-    //-------------------------------------------------------------------------
-    class UsingNullable
-    {
-        static void Main()
-        {
-            //<Snippet4>
-            int? i = 10;
-            double? d1 = 3.14;
-            bool? flag = null;
-            char? letter = 'a';
-            int?[] arr = new int?[10];
-            //</Snippet4>
-
-            // Clear compiler warnings:
-            System.Console.WriteLine(i);
-            System.Console.WriteLine(d1);
-            System.Console.WriteLine(flag);
-            System.Console.WriteLine(letter);
-
-            //<Snippet5>
-            int? x = 10;
-            if (x.HasValue)
-            {
-                System.Console.WriteLine(x.Value);
-            }
-            else
-            {
-                System.Console.WriteLine("Undefined");
-            }
-            //</Snippet5>
-
-            //<Snippet6>
-            int? y = 10;
-            if (y != null)
-            {
-                System.Console.WriteLine(y.Value);
-            }
-            else
-            {
-                System.Console.WriteLine("Undefined");
-            }
-            //</Snippet6>
-
-            //<Snippet8>
-            int? n1 = null;
-            //</Snippet8>
-            System.Console.WriteLine(n1);
-
-            //<Snippet9>
-            int? n2;
-            n2 = 10;  // Implicit conversion.
-            //</Snippet9>
-            System.Console.WriteLine(n2);
-
-            //<Snippet10>
-            int? a = 10;
-            int? b = null;
-
-            a++;         // Increment by 1, now a is 11.
-            a = a * 10;  // Multiply by 10, now a is 110.
-            a = a + b;   // Add b, now a is null.
-            //</Snippet10>
-
-            //<Snippet11>
-            int? num1 = 10;
-            int? num2 = null;
-            if (num1 >= num2)
-            {
-                Console.WriteLine("num1 is greater than or equal to num2");
-            }
-            else
-            {
-                // This clause is selected, but num1 is not less than num2.
-                Console.WriteLine("num1 >= num2 returned false (but num1 < num2 also is false)");
-            }
-
-            if (num1 < num2)
-            {
-                Console.WriteLine("num1 is less than num2");
-            }
-            else
-            {
-                // The else clause is selected again, but num1 is not greater than
-                // or equal to num2.
-                Console.WriteLine("num1 < num2 returned false (but num1 >= num2 also is false)");
-            }
-
-            if (num1 != num2)
-            {
-                // This comparison is true, num1 and num2 are not equal.
-                Console.WriteLine("Finally, num1 != num2 returns true!");
-            }
-
-            // Change the value of num1, so that both num1 and num2 are null.
-            num1 = null;
-            if (num1 == num2)
-            {
-                // The equality comparison returns true when both operands are null.
-                Console.WriteLine("num1 == num2 returns true when the value of each is null");
-            }
-
-            /* Output:
-             * num1 >= num2 returned false (but num1 < num2 also is false)
-             * num1 < num2 returned false (but num1 >= num2 also is false)
-             * Finally, num1 != num2 returns true!
-             * num1 == num2 returns true when the value of each is null
-             */
-            //</Snippet11>
-
-            //<Snippet12>
-            int? c = null;
-
-            // d = c, unless c is null, in which case d = -1.
-            int d = c ?? -1;
-            //</Snippet12>
-
-            //<Snippet13>
-            int? e = null;
-            int? f = null;
-
-            // g = e or f, unless e and f are both null, in which case g = -1.
-            int g = e ?? f ?? -1;
-            //</Snippet13>
-        }
-    }
-
-    //-------------------------------------------------------------------------
     class BoxingAndUnboxing
     {
         static void Main()
@@ -265,8 +109,43 @@ namespace CsProgGuideTypes
         }
     }
 
+    
     //-------------------------------------------------------------------------
-    //<Snippet16>
+    class BoxingValueCopy
+    {
+        static void Main()
+        {
+            //<Snippet16>
+            // Create an int variable
+            int i = 123;
+            
+            // Box the value type into an object reference
+            object o = i;  // boxing
+            
+            // Display the initial values
+            Console.WriteLine($"Value of i: {i}");
+            Console.WriteLine($"Value of boxed object o: {o}");
+            
+            // Modify the original value type
+            i = 456;
+            
+            // Display the values after modification
+            Console.WriteLine("\nAfter changing i to 456:");
+            Console.WriteLine($"Value of i: {i}");
+            Console.WriteLine($"Value of boxed object o: {o}");
+            
+    		// Output:
+    		// Value of i: 123
+    		// Value of boxed object o: 123
+    
+    		// After changing i to 456:
+    		// Value of i: 456
+    		// Value of boxed object o: 123
+            //</Snippet16>
+        }
+    }
+
+    //-------------------------------------------------------------------------
     class TestBoxing
     {
         static void Main()
@@ -279,20 +158,8 @@ namespace CsProgGuideTypes
             // Boxing copies the value of i into object o.
             object o = i;
             //</Snippet18>
-
-            // Change the value of i.
-            i = 456;
-
-            // The change in i doesn't affect the value stored in o.
-            System.Console.WriteLine("The value-type value = {0}", i);
-            System.Console.WriteLine("The object-type value = {0}", o);
         }
     }
-    /* Output:
-        The value-type value = 456
-        The object-type value = 123
-    */
-    //</Snippet16>
 
     //-------------------------------------------------------------------------
     class WrapExplicitBoxing
@@ -359,21 +226,6 @@ namespace CsProgGuideTypes
             Console.WriteLine("byte array: " + BitConverter.ToString(bytes));
             // Output: byte array: 9A-50-07-0C
             //</Snippet23>
-        }
-    }
-
-    public class ConvertStringExample4
-    {
-        static void Main(string[] args)
-        {
-            //<Snippet29>
-            string notNumber = "abc";
-            int numVal;
-            bool parsed = Int32.TryParse(notNumber, out numVal);
-
-            if (!parsed)
-                Console.WriteLine("Cannot parse '{0}' to an int.", notNumber);
-            //</Snippet29>
         }
     }
 
@@ -448,19 +300,6 @@ namespace CsProgGuideTypes
         }
     }
 
-    class ValueTypeConversion
-    {
-        static void Main()
-        {
-            //<snippet34>
-            // Implicit conversion. A long can
-            // hold any value an int can hold, and more!
-            int num = 2147483647;
-            long bigNum = num;
-            //</snippet34>
-        }
-    }
-
     class ConvertHexStrings2
     {
         static void Main()
@@ -507,193 +346,4 @@ namespace CsProgGuideTypes
             System.Console.ReadKey();
         }
     }
-
-    //<snippet40>
-    class SafeCasting
-    {
-        class Animal
-        {
-            public void Eat() { Console.WriteLine("Eating."); }
-            public override string ToString()
-            {
-                return "I am an animal.";
-            }
-        }
-        class Mammal : Animal { }
-        class Giraffe : Mammal { }
-
-        class SuperNova { }
-
-        static void Main()
-        {
-            SafeCasting app = new SafeCasting();
-
-            // Use the is operator to verify the type.
-            // before performing a cast.
-            Giraffe g = new Giraffe();
-            app.UseIsOperator(g);
-
-            // Use the as operator and test for null
-            // before referencing the variable.
-            app.UseAsOperator(g);
-
-            // Use the as operator to test
-            // an incompatible type.
-            SuperNova sn = new SuperNova();
-            app.UseAsOperator(sn);
-
-            // Use the as operator with a value type.
-            // Note the implicit conversion to int? in
-            // the method body.
-            int i = 5;
-            app.UseAsWithNullable(i);
-
-            double d = 9.78654;
-            app.UseAsWithNullable(d);
-
-            // Keep the console window open in debug mode.
-            System.Console.WriteLine("Press any key to exit.");
-            System.Console.ReadKey();
-        }
-
-        void UseIsOperator(Animal a)
-        {
-            if (a is Mammal)
-            {
-                Mammal m = (Mammal)a;
-                m.Eat();
-            }
-        }
-
-        void UseAsOperator(object o)
-        {
-            Mammal? m = o as Mammal;
-            if (m != null)
-            {
-                Console.WriteLine(m.ToString());
-            }
-            else
-            {
-                Console.WriteLine("{0} is not a Mammal", o.GetType().Name);
-            }
-        }
-
-        void UseAsWithNullable(System.ValueType val)
-        {
-            int? j = val as int?;
-            if (j != null)
-            {
-                Console.WriteLine(j);
-            }
-            else
-            {
-                Console.WriteLine("Could not convert " + val.ToString());
-            }
-        }
-    }
-    //</snippet40>
-
-    //<snippet41>
-    class Animal
-    {
-        public void Eat() => System.Console.WriteLine("Eating.");
-
-        public override string ToString() => "I am an animal.";
-    }
-
-    class Reptile : Animal { }
-    class Mammal : Animal { }
-
-    class UnSafeCast
-    {
-        static void Main()
-        {
-            Test(new Mammal());
-
-            // Keep the console window open in debug mode.
-            System.Console.WriteLine("Press any key to exit.");
-            System.Console.ReadKey();
-        }
-
-        static void Test(Animal a)
-        {
-            // System.InvalidCastException at run time
-            // Unable to cast object of type 'Mammal' to type 'Reptile'
-            Reptile r = (Reptile)a;
-        }
-    }
-    //</snippet41>
-
-    //<snippet43>
-    class ValueTypeDemo
-    {
-        static void PassByValue(int i)
-        {
-
-            i = 5;
-            Console.WriteLine("PassByValue: local copy = {0}", i);
-        }
-
-        static void PassByReference(ref int i)
-        {
-
-            i = 5;
-            Console.WriteLine("PassByReference: local copy = {0}", i);
-        }
-
-        static void Main()
-        {
-            int num = 1;
-
-            // Pass num by value (default behavior for value types).
-            PassByValue(num);
-            Console.WriteLine("After calling PassByValue, num = {0}", num);
-
-            //Pass num by reference.
-            PassByReference(ref num);
-            Console.WriteLine("After calling PassByReference, num = {0}", num);
-
-            // Keep the console open in debug mode.
-            Console.WriteLine("Press any key to exit.");
-            Console.ReadKey();
-        }
-    }
-    /*
-       Output:
-        PassByValue: local copy = 5
-        After calling PassByValue, num = 1
-        PassByReference: local copy = 5
-        After calling PassByReference, num = 5
-    */
-    //</snippet43>
-
-    //<Snippet46>
-    class Base { }
-    class Derived : Base { }
-
-    class GetTypeInfo
-    {
-
-        static void Main()
-        {
-            Derived derived = new Derived();
-
-            // The 'is' keyword is polymorphic.
-
-            Console.WriteLine("derived is Derived = {0}", derived is Derived);
-            // true:
-            Console.WriteLine("derived is Base = {0}", derived is Base);
-
-            // GetType returns the exact run-time type.
-            object o = derived;
-            Console.WriteLine("The run-time type of o is {0}", o.GetType().Name);
-        }
-    }
-    /* Output:
-        derived is Derived = True
-        derived is Base = True
-        The run-time type of o is Derived
-     */
-
-    //</Snippet46>
 }
