@@ -1,32 +1,30 @@
-﻿Imports Microsoft.Win32.SafeHandles
-Imports System.Runtime.InteropServices
+﻿Imports System.IO
 
-Public Class BaseClassWithSafeHandle
+Public Class DisposableBase
     Implements IDisposable
 
-    ' To detect redundant calls
-    Private _disposedValue As Boolean
+    ' Detect redundant Dispose() calls.
+    Private _isDisposed As Boolean
 
-    ' Instantiate a SafeHandle instance.
-    Private _safeHandle As SafeHandle = New SafeFileHandle(IntPtr.Zero, True)
+    ' Instantiate a disposable object owned by this class.
+    Private _managedResource As Stream = New MemoryStream()
 
     ' Public implementation of Dispose pattern callable by consumers.
-    Public Sub Dispose() _
-               Implements IDisposable.Dispose
+    Public Sub Dispose() Implements IDisposable.Dispose
         Dispose(True)
         GC.SuppressFinalize(Me)
     End Sub
 
     ' Protected implementation of Dispose pattern.
-    Protected Overridable Sub Dispose(ByVal disposing As Boolean)
-        If Not _disposedValue Then
+    Protected Overridable Sub Dispose(disposing As Boolean)
+        If Not _isDisposed Then
+            _isDisposed = True
 
             If disposing Then
-                _safeHandle?.Dispose()
-                _safeHandle = Nothing
+                ' Dispose managed state.
+                _managedResource?.Dispose()
+                _managedResource = Nothing
             End If
-
-            _disposedValue = True
         End If
     End Sub
 End Class
