@@ -1,31 +1,43 @@
-﻿Public Class BaseClassWithFinalizer
+﻿Imports System
+Imports System.IO
+Imports System.Runtime.InteropServices
+Imports System.Threading
+
+Public Class DisposableBaseWithFinalizer
     Implements IDisposable
 
-    ' To detect redundant calls
-    Private _disposedValue As Boolean
+    ' Detect redundant Dispose() calls in a thread-safe manner.
+    ' _isDisposed == 0 means Dispose(bool) has not been called yet.
+    ' _isDisposed == 1 means Dispose(bool) has been already called.
+    Private _isDisposed As Integer
+
+    ' Instantiate a disposable object owned by this class.
+    Private _managedResource As Stream = New MemoryStream()
+
+    ' A pointer to 10 bytes allocated on the unmanaged heap.
+    Private _unmanagedResource As IntPtr = Marshal.AllocHGlobal(10)
 
     Protected Overrides Sub Finalize()
         Dispose(False)
     End Sub
 
     ' Public implementation of Dispose pattern callable by consumers.
-    Public Sub Dispose() _
-               Implements IDisposable.Dispose
+    Public Sub Dispose() Implements IDisposable.Dispose
         Dispose(True)
         GC.SuppressFinalize(Me)
     End Sub
 
     ' Protected implementation of Dispose pattern.
-    Protected Overridable Sub Dispose(ByVal disposing As Boolean)
-        If Not _disposedValue Then
-
+    Protected Overridable Sub Dispose(disposing As Boolean)
+        ' In case _isDisposed is 0, atomically set it to 1.
+        ' Enter the branch only if the original value is 0.
+        If Interlocked.CompareExchange(_isDisposed, 1, 0) = 0 Then
             If disposing Then
-                ' TODO: dispose managed state (managed objects)
+                _managedResource?.Dispose()
+                _managedResource = Nothing
             End If
 
-            ' TODO free unmanaged resources (unmanaged objects) And override finalizer
-            ' TODO: set large fields to null
-            _disposedValue = True
+            Marshal.FreeHGlobal(_unmanagedResource)
         End If
     End Sub
 End Class
